# Improve Windows test build performance by building projects in groups - 4x less memory 326% faster build.