Fixes #14594 and is a part of #17149 

The build time improvement for priority 1 tests: 02:09:09.700 (2 hours 9 minutes 9 seconds .700) on 03:46:42.325 original time and now takes 01:37:32.619 what means that build and test is **~132% faster**.

Test build time improved from 02:45:09:518 to 00:38:47.199 what is **~326% faster.**