@4creators 2018-03-23
Performance comparison between priority 1 build and test runs - time elapsed from build start is on the left:

With MSBuild cache workaround:
https://ci.dot.net/job/dotnet_coreclr/job/master/job/jitstress/job/x64_checked_windows_nt_jitincompletehwintrinsic_prtest/231/
```
00:15:46.973 BUILDTEST: Commencing CoreCLR repo test build
00:15:46.976 BUILDTEST: Checking prerequisites
```
```
00:54:34.172 BUILDTEST: Test build successful.
00:54:34.172 BUILDTEST: Test binaries are available at D:\j\workspace\x64_checked_w---1cba8152\bin\tests\Windows_NT.x64.Checked
```
```
01:37:31.510   === TEST EXECUTION SUMMARY ===
01:37:31.517      baseservices.compilerservices.XUnitWrapper       Total:    10, Errors: 0, Failed: 0, Skipped: 0, Time:    12.419s
01:37:31.517      baseservices.exceptions.XUnitWrapper             Total:   139, Errors: 0, Failed: 0, Skipped: 0, Time:   547.453s
01:37:31.517      baseservices.multidimmarray.XUnitWrapper         Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     7.898s
01:37:31.517      baseservices.regression.XUnitWrapper             Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     6.425s
01:37:31.517      baseservices.threading.XUnitWrapper              Total:   433, Errors: 0, Failed: 0, Skipped: 0, Time:  2106.365s
01:37:31.517      baseservices.varargs.XUnitWrapper                Total:     2, Errors: 0, Failed: 0, Skipped: 0, Time:     6.255s
01:37:31.517      CoreMangLib.components.XUnitWrapper              Total:     2, Errors: 0, Failed: 0, Skipped: 0, Time:     5.843s
01:37:31.517      CoreMangLib.cti.XUnitWrapper                     Total:  2055, Errors: 0, Failed: 0, Skipped: 0, Time:  2506.745s
01:37:31.517      CoreMangLib.system.XUnitWrapper                  Total:    16, Errors: 0, Failed: 0, Skipped: 0, Time:    77.553s
01:37:31.517      Exceptions.Finalization.XUnitWrapper             Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     6.111s
01:37:31.517      Exceptions.ForeignThread.XUnitWrapper            Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     6.342s
01:37:31.517      GC.API.XUnitWrapper                              Total:    71, Errors: 0, Failed: 0, Skipped: 0, Time:   366.584s
01:37:31.517      GC.Coverage.XUnitWrapper                         Total:     3, Errors: 0, Failed: 0, Skipped: 0, Time:    28.225s
01:37:31.517      GC.Features.XUnitWrapper                         Total:    41, Errors: 0, Failed: 0, Skipped: 0, Time:   224.396s
01:37:31.517      GC.LargeMemory.XUnitWrapper                      Total:     6, Errors: 0, Failed: 0, Skipped: 0, Time:    19.568s
01:37:31.517      GC.Regressions.XUnitWrapper                      Total:    11, Errors: 0, Failed: 0, Skipped: 0, Time:    45.906s
01:37:31.517      GC.Scenarios.XUnitWrapper                        Total:   479, Errors: 0, Failed: 0, Skipped: 0, Time:  1006.173s
01:37:31.517      Interop.ArrayMarshalling.XUnitWrapper            Total:     2, Errors: 0, Failed: 0, Skipped: 0, Time:    33.890s
01:37:31.517      Interop.BestFitMapping.XUnitWrapper              Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    13.124s
01:37:31.517      Interop.FuncPtrAsDelegateParam.XUnitWrapper      Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    14.027s
01:37:31.517      Interop.ICastable.XUnitWrapper                   Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    12.461s
01:37:31.517      Interop.MarshalAPI.XUnitWrapper                  Total:    26, Errors: 0, Failed: 0, Skipped: 0, Time:    55.151s
01:37:31.517      Interop.NativeCallable.XUnitWrapper              Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    11.106s
01:37:31.517      Interop.PrimitiveMarshalling.XUnitWrapper        Total:     3, Errors: 0, Failed: 0, Skipped: 0, Time:    11.373s
01:37:31.517      Interop.RefCharArray.XUnitWrapper                Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     4.707s
01:37:31.517      Interop.RefInt.XUnitWrapper                      Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    13.005s
01:37:31.517      Interop.SimpleStruct.XUnitWrapper                Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     8.005s
01:37:31.517      Interop.SizeConst.XUnitWrapper                   Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     6.418s
01:37:31.517      Interop.StringMarshalling.XUnitWrapper           Total:     3, Errors: 0, Failed: 0, Skipped: 0, Time:     6.178s
01:37:31.517      Interop.StructMarshalling.XUnitWrapper           Total:     2, Errors: 0, Failed: 0, Skipped: 0, Time:    21.847s
01:37:31.517      Interop.StructPacking.XUnitWrapper               Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    16.254s
01:37:31.517      JIT.BBT.XUnitWrapper                             Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    10.066s
01:37:31.517      JIT.CheckProjects.XUnitWrapper                   Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    30.279s
01:37:31.517      JIT.CodeGenBringUpTests.XUnitWrapper             Total:   159, Errors: 0, Failed: 0, Skipped: 0, Time:   582.949s
01:37:31.517      JIT.Directed.XUnitWrapper                        Total:   607, Errors: 0, Failed: 0, Skipped: 0, Time:  1338.978s
01:37:31.517      JIT.Generics.XUnitWrapper                        Total:   219, Errors: 0, Failed: 0, Skipped: 0, Time:   740.878s
01:37:31.517      JIT.HardwareIntrinsics.XUnitWrapper              Total:   355, Errors: 0, Failed: 0, Skipped: 0, Time:   948.049s
01:37:31.517      JIT.IL_Conformance.XUnitWrapper                  Total:   404, Errors: 0, Failed: 0, Skipped: 0, Time:   938.332s
01:37:31.517      JIT.Intrinsics.XUnitWrapper                      Total:    12, Errors: 0, Failed: 0, Skipped: 0, Time:    79.735s
01:37:31.517      JIT.jit64.XUnitWrapper                           Total:   865, Errors: 0, Failed: 0, Skipped: 0, Time:  1726.639s
01:37:31.518      JIT.Methodical.XUnitWrapper                      Total:  2075, Errors: 0, Failed: 0, Skipped: 0, Time:  2425.437s
01:37:31.518      JIT.opt.XUnitWrapper                             Total:   137, Errors: 0, Failed: 0, Skipped: 0, Time:   471.389s
01:37:31.518      JIT.Performance.XUnitWrapper                     Total:    83, Errors: 0, Failed: 0, Skipped: 0, Time:   434.781s
01:37:31.518      JIT.Regression.XUnitWrapper                      Total:  1275, Errors: 0, Failed: 0, Skipped: 0, Time:  1802.919s
01:37:31.518      JIT.RyuJIT.XUnitWrapper                          Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     5.931s
01:37:31.518      JIT.SIMD.XUnitWrapper                            Total:   108, Errors: 0, Failed: 0, Skipped: 0, Time:   424.571s
01:37:31.518      JIT.superpmi.XUnitWrapper                        Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:   235.119s
01:37:31.518      Loader.binding.XUnitWrapper                      Total:     7, Errors: 0, Failed: 0, Skipped: 0, Time:   234.411s
01:37:31.518      Loader.classloader.XUnitWrapper                  Total:  1973, Errors: 0, Failed: 0, Skipped: 0, Time:  2388.132s
01:37:31.518      Loader.lowlevel.XUnitWrapper                     Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    32.062s
01:37:31.518      Loader.multimodule.XUnitWrapper                  Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    28.480s
01:37:31.518      Loader.NativeLibs.XUnitWrapper                   Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     7.888s
01:37:31.518      Loader.regressions.XUnitWrapper                  Total:     4, Errors: 0, Failed: 0, Skipped: 0, Time:    39.815s
01:37:31.518      Loader.versioning.XUnitWrapper                   Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    17.893s
01:37:31.518      managed.Compilation.XUnitWrapper                 Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    51.448s
01:37:31.518      readytorun.tests.XUnitWrapper                    Total:     6, Errors: 0, Failed: 0, Skipped: 0, Time:    39.477s
01:37:31.518      reflection.DefaultInterfaceMethods.XUnitWrapper  Total:     3, Errors: 0, Failed: 0, Skipped: 0, Time:    30.021s
01:37:31.518      reflection.ldtoken.XUnitWrapper                  Total:     2, Errors: 0, Failed: 0, Skipped: 0, Time:    35.206s
01:37:31.518      reflection.regression.XUnitWrapper               Total:     2, Errors: 0, Failed: 0, Skipped: 0, Time:    29.719s
01:37:31.518      reflection.Tier1Collectible.XUnitWrapper         Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    25.760s
01:37:31.518      Regressions.assemblyref.XUnitWrapper             Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     7.443s
01:37:31.518      Regressions.common.XUnitWrapper                  Total:    15, Errors: 0, Failed: 0, Skipped: 0, Time:    83.409s
01:37:31.518      Regressions.coreclr.XUnitWrapper                 Total:    53, Errors: 0, Failed: 0, Skipped: 0, Time:   347.476s
01:37:31.518      Regressions.expl_double.XUnitWrapper             Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    10.362s
01:37:31.518      tracing.eventactivityidcontrol.XUnitWrapper      Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    10.253s
01:37:31.518      tracing.eventlistener.XUnitWrapper               Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    10.980s
01:37:31.518      tracing.eventsourcetrace.XUnitWrapper.dll        Total:     0
01:37:31.518      tracing.tracevalidation.XUnitWrapper             Total:     3, Errors: 0, Failed: 0, Skipped: 0, Time:    28.782s
01:37:31.518                                                              -----          -          -           -        ----------
01:37:31.519                                                 GRAND TOTAL: 11698          0          0           0        22862.879s (2553.701s)
01:37:32.619 
01:37:32.619 Build succeeded.
01:37:32.619     0 Warning(s)
01:37:32.619     0 Error(s)
01:37:32.619 
01:37:32.619 Time Elapsed 00:42:36.95
01:37:32.625 RUNTEST: Test run successful. Refer to the log files for details:
01:37:32.625     D:\j\workspace\x64_checked_w---1cba8152\tests\..\bin\Logs\TestRun_Windows_NT__x64__Checked.html
01:37:32.625     D:\j\workspace\x64_checked_w---1cba8152\tests\..\bin\Logs\TestRun_Windows_NT__x64__Checked.xml
01:37:32.976 [x64_checked_w---1cba8152] $ cmd /c call C:\Users\DOTNET~1\AppData\Local\Temp\jenkins6498336450786809472.bat
01:37:32.988 
01:37:32.988 D:\j\workspace\x64_checked_w---1cba8152>build.cmd checked x64 linuxmscorlib 
```

Without MSBuild cache workaround:
https://ci.dot.net/job/dotnet_coreclr/job/master/job/jitstress/job/x64_checked_windows_nt_jitincompletehwintrinsic_prtest/230/
```
00:16:12.482 BUILDTEST: Commencing CoreCLR repo test build
00:16:12.485 BUILDTEST: Checking prerequisites
```

```
03:01:22.000 BUILDTEST: Test build successful.
03:01:22.000 BUILDTEST: Test binaries are available at D:\j\workspace\x64_checked_w---1cba8152\bin\tests\Windows_NT.x64.Checked
```
```
03:46:41.118   === TEST EXECUTION SUMMARY ===
03:46:41.123      baseservices.compilerservices.XUnitWrapper       Total:    10, Errors: 0, Failed: 0, Skipped: 0, Time:    14.889s
03:46:41.123      baseservices.exceptions.XUnitWrapper             Total:   139, Errors: 0, Failed: 0, Skipped: 0, Time:   556.110s
03:46:41.123      baseservices.multidimmarray.XUnitWrapper         Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     8.387s
03:46:41.124      baseservices.regression.XUnitWrapper             Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     8.351s
03:46:41.124      baseservices.threading.XUnitWrapper              Total:   433, Errors: 0, Failed: 0, Skipped: 0, Time:  2186.194s
03:46:41.124      baseservices.varargs.XUnitWrapper                Total:     2, Errors: 0, Failed: 0, Skipped: 0, Time:     6.183s
03:46:41.124      CoreMangLib.components.XUnitWrapper              Total:     2, Errors: 0, Failed: 0, Skipped: 0, Time:     4.909s
03:46:41.127      CoreMangLib.cti.XUnitWrapper                     Total:  2055, Errors: 0, Failed: 0, Skipped: 0, Time:  2564.404s
03:46:41.127      CoreMangLib.system.XUnitWrapper                  Total:    16, Errors: 0, Failed: 0, Skipped: 0, Time:    69.250s
03:46:41.127      Exceptions.Finalization.XUnitWrapper             Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     4.864s
03:46:41.127      Exceptions.ForeignThread.XUnitWrapper            Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     9.009s
03:46:41.127      GC.API.XUnitWrapper                              Total:    71, Errors: 0, Failed: 0, Skipped: 0, Time:   372.295s
03:46:41.127      GC.Coverage.XUnitWrapper                         Total:     3, Errors: 0, Failed: 0, Skipped: 0, Time:    18.237s
03:46:41.127      GC.Features.XUnitWrapper                         Total:    41, Errors: 0, Failed: 0, Skipped: 0, Time:   247.814s
03:46:41.128      GC.LargeMemory.XUnitWrapper                      Total:     6, Errors: 0, Failed: 0, Skipped: 0, Time:    37.409s
03:46:41.128      GC.Regressions.XUnitWrapper                      Total:    11, Errors: 0, Failed: 0, Skipped: 0, Time:    37.986s
03:46:41.128      GC.Scenarios.XUnitWrapper                        Total:   479, Errors: 0, Failed: 0, Skipped: 0, Time:  1093.970s
03:46:41.128      hosting.stress.XUnitWrapper                      Total:     3, Errors: 0, Failed: 0, Skipped: 0, Time:    24.110s
03:46:41.128      Interop.ArrayMarshalling.XUnitWrapper            Total:     2, Errors: 0, Failed: 0, Skipped: 0, Time:    23.882s
03:46:41.129      Interop.BestFitMapping.XUnitWrapper              Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    19.751s
03:46:41.129      Interop.FuncPtrAsDelegateParam.XUnitWrapper      Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    20.609s
03:46:41.129      Interop.ICastable.XUnitWrapper                   Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    20.610s
03:46:41.129      Interop.MarshalAPI.XUnitWrapper                  Total:    26, Errors: 0, Failed: 0, Skipped: 0, Time:    56.441s
03:46:41.130      Interop.NativeCallable.XUnitWrapper              Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     5.806s
03:46:41.130      Interop.PrimitiveMarshalling.XUnitWrapper        Total:     3, Errors: 0, Failed: 0, Skipped: 0, Time:    11.155s
03:46:41.130      Interop.RefCharArray.XUnitWrapper                Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     9.375s
03:46:41.130      Interop.RefInt.XUnitWrapper                      Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     8.557s
03:46:41.130      Interop.SimpleStruct.XUnitWrapper                Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     8.115s
03:46:41.130      Interop.SizeConst.XUnitWrapper                   Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     8.139s
03:46:41.130      Interop.StringMarshalling.XUnitWrapper           Total:     3, Errors: 0, Failed: 0, Skipped: 0, Time:     9.115s
03:46:41.130      Interop.StructMarshalling.XUnitWrapper           Total:     2, Errors: 0, Failed: 0, Skipped: 0, Time:     9.275s
03:46:41.130      Interop.StructPacking.XUnitWrapper               Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     6.358s
03:46:41.130      JIT.BBT.XUnitWrapper                             Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     5.285s
03:46:41.130      JIT.CheckProjects.XUnitWrapper                   Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    61.803s
03:46:41.130      JIT.CodeGenBringUpTests.XUnitWrapper             Total:   159, Errors: 0, Failed: 0, Skipped: 0, Time:   590.790s
03:46:41.130      JIT.Directed.XUnitWrapper                        Total:   607, Errors: 0, Failed: 0, Skipped: 0, Time:  1380.334s
03:46:41.130      JIT.Generics.XUnitWrapper                        Total:   219, Errors: 0, Failed: 0, Skipped: 0, Time:   789.397s
03:46:41.130      JIT.HardwareIntrinsics.XUnitWrapper              Total:   355, Errors: 0, Failed: 0, Skipped: 0, Time:   962.908s
03:46:41.130      JIT.IL_Conformance.XUnitWrapper                  Total:   404, Errors: 0, Failed: 0, Skipped: 0, Time:  1009.794s
03:46:41.130      JIT.Intrinsics.XUnitWrapper                      Total:    12, Errors: 0, Failed: 0, Skipped: 0, Time:   111.961s
03:46:41.130      JIT.jit64.XUnitWrapper                           Total:   865, Errors: 0, Failed: 0, Skipped: 0, Time:  1643.230s
03:46:41.130      JIT.Methodical.XUnitWrapper                      Total:  2075, Errors: 0, Failed: 0, Skipped: 0, Time:  2498.825s
03:46:41.131      JIT.opt.XUnitWrapper                             Total:   137, Errors: 0, Failed: 0, Skipped: 0, Time:   608.545s
03:46:41.131      JIT.Performance.XUnitWrapper                     Total:    83, Errors: 0, Failed: 0, Skipped: 0, Time:   446.113s
03:46:41.131      JIT.Regression.XUnitWrapper                      Total:  1275, Errors: 0, Failed: 0, Skipped: 0, Time:  2067.152s
03:46:41.131      JIT.RyuJIT.XUnitWrapper                          Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    21.914s
03:46:41.131      JIT.SIMD.XUnitWrapper                            Total:   108, Errors: 0, Failed: 0, Skipped: 0, Time:   413.576s
03:46:41.131      JIT.superpmi.XUnitWrapper                        Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:   244.082s
03:46:41.131      Loader.binding.XUnitWrapper                      Total:     7, Errors: 0, Failed: 0, Skipped: 0, Time:    47.518s
03:46:41.131      Loader.classloader.XUnitWrapper                  Total:  1973, Errors: 0, Failed: 0, Skipped: 0, Time:  2521.864s
03:46:41.131      Loader.lowlevel.XUnitWrapper                     Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    32.582s
03:46:41.131      Loader.multimodule.XUnitWrapper                  Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    24.523s
03:46:41.131      Loader.NativeLibs.XUnitWrapper                   Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    25.955s
03:46:41.131      Loader.regressions.XUnitWrapper                  Total:     4, Errors: 0, Failed: 0, Skipped: 0, Time:    24.066s
03:46:41.133      Loader.versioning.XUnitWrapper                   Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    20.336s
03:46:41.133      managed.Compilation.XUnitWrapper                 Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    55.609s
03:46:41.133      readytorun.tests.XUnitWrapper                    Total:     6, Errors: 0, Failed: 0, Skipped: 0, Time:    41.859s
03:46:41.133      reflection.DefaultInterfaceMethods.XUnitWrapper  Total:     3, Errors: 0, Failed: 0, Skipped: 0, Time:    35.479s
03:46:41.133      reflection.ldtoken.XUnitWrapper                  Total:     2, Errors: 0, Failed: 0, Skipped: 0, Time:    13.232s
03:46:41.133      reflection.regression.XUnitWrapper               Total:     2, Errors: 0, Failed: 0, Skipped: 0, Time:    13.739s
03:46:41.133      reflection.Tier1Collectible.XUnitWrapper         Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    46.888s
03:46:41.133      Regressions.assemblyref.XUnitWrapper             Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:     9.073s
03:46:41.133      Regressions.common.XUnitWrapper                  Total:    15, Errors: 0, Failed: 0, Skipped: 0, Time:    90.268s
03:46:41.133      Regressions.coreclr.XUnitWrapper                 Total:    53, Errors: 0, Failed: 0, Skipped: 0, Time:   245.127s
03:46:41.134      Regressions.expl_double.XUnitWrapper             Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    16.704s
03:46:41.134      tracing.eventactivityidcontrol.XUnitWrapper      Total:     1, Errors: 0, Failed: 0, Skipped: 0, Time:    44.236s
…